+++ .example.env
@@ -6,3 +6,7 @@ ERC4337_ACCOUNT_FACTORY='0x123'
 ERC4337_ORIGIN_HEADER='x'
 API_KEY='x'
 SENTRY_URL='x'
+PINATA_BASE_URL='x'
+PINATA_API_KEY='x'
+PINATA_SECRET_API_KEY='x'
+IPFS_URL='x'
